refactor : separate configuration files into new package
