+++ src/main/java/com/ironknee/Knee2KneelSpring/configuration/CorsMvcConfig.java
@@ -1,4 +1,4 @@
-package com.ironknee.Knee2KneelSpring.authentication;
+package com.ironknee.Knee2KneelSpring.configuration;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
